type refinement + change env supabase url for docs
