+++ frontend/src/types/class.ts
@@ -36,3 +36,32 @@ export interface CreateClassData {
 export interface UpdateClassData extends CreateClassData {
   id: number;
 }
+
+export interface NextPresentation {
+  id: number;
+  child_id: number;
+  class_id: number;
+  name: string;
+  category?: string;
+  status: string;
+  notes?: string;
+  class_name: string;
+  child_firstname: string;
+  child_surname: string;
+  created_by_firstname?: string;
+  created_by_surname?: string;
+  updated_by_firstname?: string;
+  updated_by_surname?: string;
+}
+
+export interface ClassAttendanceRow {
+  id: number;
+  firstname: string;
+  surname: string;
+  attendance_date: string | null;
+  check_in_at: string | null;
+  check_out_at: string | null;
+  checked_in_by: number | null;
+  checked_out_by: number | null;
+  notes: string | null;
+}
